// packages/client/ui/src/components/organisms/data-table/types.ts
type DataKey<TData> = Extract<keyof TData, string> | string

export type DataTableRowAction<TData> = {
  label: string
  onSelect: (row: TData) => void
  id?: string
  destructive?: boolean
  disabled?: (row: TData) => boolean
}

type BaseColumn<TData> = {
  header: string
  accessorKey?: DataKey<TData>
  enableSorting?: boolean
  enableHiding?: boolean
}
// ...
export type DataTableNumberColumn<TData> = BaseColumn<TData> & {
  type: 'number'
  accessorKey: DataKey<TData>
  format?: (value: number, row: TData) => React.ReactNode
  locale?: string
  minimumFractionDigits?: number
  maximumFractionDigits?: number
}